engine:core: Create resource manager
The resource manager can keep track of singleton like resources of type
T.

When we add a resource and it is not already registered we initialize it
with its default value.

If the resource type is already registered we get an error.

We are able to also remove the resources.

We are also able to mutate the resource.

Close #5