First GhostEngine application
Integrated both the resource manager and the ECS into the application.

We give a basic scaffolding for an application in the sandbox project.

At the moment we have a forwarding layer that forwards methods calls
from the application to either the resource manager or the ECS.

This API is not stable yet, but this one seems the way. We shall see how
it feels.

Closes #14.